+++ driver-core/src/main/java/com/datastax/driver/core/ControlConnection.java
@@ -411,7 +411,7 @@ class ControlConnection implements Host.StateListener {
             return null;
         } else if (addr == null) {
             if (logMissingRpcAddresses)
-                logger.error("No rpc_address found for host {} in {}'s peers system table. {} will be ignored.", peer, connectedHost, peer);
+                logger.warn("No rpc_address found for host {} in {}'s peers system table. {} will be ignored.", peer, connectedHost, peer);
             return null;
         } else if (addr.equals(bindAllAddress)) {
             logger.warn("Found host with 0.0.0.0 as rpc_address, using listen_address ({}) to contact it instead. If this is incorrect you should avoid the use of 0.0.0.0 server side.", peer);
@@ -461,18 +461,18 @@ class ControlConnection implements Host.StateListener {
                     // until the control connection is back up (which leads to a catch-22 if there is only one)
                     return true;
                 } else {
-                    logger.error("No row found for host {} in {}'s peers system table. {} will be ignored.", host.getAddress(), c.address, host.getAddress());
+                    logger.warn("No row found for host {} in {}'s peers system table. {} will be ignored.", host.getAddress(), c.address, host.getAddress());
                     return false;
                 }
                 // Ignore hosts with a null rpc_address, as this is most likely a phantom row in system.peers (JAVA-428).
                 // Don't test this for the control host since we're already connected to it anyway, and we read the info from system.local
                 // which doesn't have an rpc_address column (JAVA-546).
             } else if (!c.address.equals(host.getSocketAddress()) && row.getInet("rpc_address") == null) {
-                logger.error("No rpc_address found for host {} in {}'s peers system table. {} will be ignored.", host.getAddress(), c.address, host.getAddress());
+                logger.warn("No rpc_address found for host {} in {}'s peers system table. {} will be ignored.", host.getAddress(), c.address, host.getAddress());
                 return false;
             }
 
-            updateInfo(host, row, cluster);
+            updateInfo(host, row, cluster, false);
             return true;
 
         } catch (ConnectionException e) {
@@ -499,9 +499,9 @@ class ControlConnection implements Host.StateListener {
     }
 
     // row can come either from the 'local' table or the 'peers' one
-    private static void updateInfo(Host host, Row row, Cluster.Manager cluster) {
+    private static void updateInfo(Host host, Row row, Cluster.Manager cluster, boolean isInitialConnection) {
         if (!row.isNull("data_center") || !row.isNull("rack"))
-            updateLocationInfo(host, row.getString("data_center"), row.getString("rack"), cluster);
+            updateLocationInfo(host, row.getString("data_center"), row.getString("rack"), isInitialConnection, cluster);
 
         String version = row.getString("release_version");
         // We don't know if it's a 'local' or a 'peers' row, and only 'peers' rows have the 'peer' field.
@@ -512,17 +512,17 @@ class ControlConnection implements Host.StateListener {
         host.setVersionAndListenAdress(version, listenAddress);
     }
 
-    private static void updateLocationInfo(Host host, String datacenter, String rack, Cluster.Manager cluster) {
+    private static void updateLocationInfo(Host host, String datacenter, String rack, boolean isInitialConnection, Cluster.Manager cluster) {
         if (Objects.equal(host.getDatacenter(), datacenter) && Objects.equal(host.getRack(), rack))
             return;
 
         // If the dc/rack information changes for an existing node, we need to update the load balancing policy.
         // For that, we remove and re-add the node against the policy. Not the most elegant, and assumes
         // that the policy will update correctly, but in practice this should work.
-        if (!host.wasJustAdded())
+        if (!isInitialConnection)
             cluster.loadBalancingPolicy().onDown(host);
         host.setLocationInfo(datacenter, rack);
-        if (!host.wasJustAdded())
+        if (!isInitialConnection)
             cluster.loadBalancingPolicy().onAdd(host);
     }
 
@@ -556,7 +556,7 @@ class ControlConnection implements Host.StateListener {
             if (host == null) {
                 logger.debug("Host in local system table ({}) unknown to us (ok if said host just got removed)", connection.address);
             } else {
-                updateInfo(host, localRow, cluster);
+                updateInfo(host, localRow, cluster, isInitialConnection);
                 Set<String> tokens = localRow.getSet("tokens", String.class);
                 if (partitioner != null && !tokens.isEmpty())
                     tokenMap.put(host, tokens);
@@ -593,7 +593,7 @@ class ControlConnection implements Host.StateListener {
                 isNew = true;
             }
             if (dcs.get(i) != null || racks.get(i) != null)
-                updateLocationInfo(host, dcs.get(i), racks.get(i), cluster);
+                updateLocationInfo(host, dcs.get(i), racks.get(i), isInitialConnection, cluster);
             if (cassandraVersions.get(i) != null)
                 host.setVersionAndListenAdress(cassandraVersions.get(i), listenAddresses.get(i));