+++ driver-core/src/main/java/com/datastax/driver/core/DynamicConnectionPool.java
@@ -250,6 +250,8 @@ class DynamicConnectionPool extends HostConnectionPool {
 
     @Override
     public void returnConnection(PooledConnection connection) {
+        int inFlight = connection.inFlight.decrementAndGet();
+
         if (isClosed()) {
             close(connection);
             return;
@@ -260,8 +262,6 @@ class DynamicConnectionPool extends HostConnectionPool {
             // closed the pool.
             return;
         }
-
-        int inFlight = connection.inFlight.decrementAndGet();
 
         if (trash.contains(connection)) {
             if (inFlight == 0 && trash.remove(connection))