+++ driver-core/src/main/java/com/datastax/driver/core/RequestHandler.java
@@ -67,7 +67,6 @@ class RequestHandler implements Connection.ResponseCallback {
 
     private volatile Map<InetSocketAddress, Throwable> errors;
 
-    private volatile boolean isCanceled;
     private volatile Connection.ResponseHandler connectionHandler;
 
     private final Timer.Context timerContext;
@@ -99,7 +98,7 @@ class RequestHandler implements Connection.ResponseCallback {
 
     public void sendRequest() {
         try {
-            while (queryPlan.hasNext() && !isCanceled) {
+            while (queryPlan.hasNext() && !queryStateRef.get().isCancelled()) {
                 Host host = queryPlan.next();
                 logger.trace("Querying node {}", host);
                 if (query(host))
@@ -126,7 +125,7 @@ class RequestHandler implements Connection.ResponseCallback {
                 triedHosts.add(current);
             }
             current = host;
-            connectionHandler = connection.write(this);
+            write(connection, this);
             return true;
         } catch (ConnectionException e) {
             // If we have any problem with the connection, move to the next node.
@@ -154,6 +153,38 @@ class RequestHandler implements Connection.ResponseCallback {
         }
     }
 
+    private void write(Connection connection, Connection.ResponseCallback responseCallback) throws ConnectionException, BusyConnectionException {
+        // Make sure cancel() does not see a stale connectionHandler if it sees the new query state
+        // before connection.write has completed
+        connectionHandler = null;
+
+        // Ensure query state is "in progress" (can be already if connection.write failed on a previous node and we're retrying)
+        while (true) {
+            QueryState previous = queryStateRef.get();
+            if (previous.isCancelled()) {
+                if (connection instanceof PooledConnection)
+                    ((PooledConnection)connection).release();
+                return;
+            }
+            if (previous.inProgress || queryStateRef.compareAndSet(previous, previous.startNext()))
+                break;
+        }
+
+        connectionHandler = connection.write(responseCallback, false);
+        // Only start the timeout when we're sure connectionHandler is set. This avoids an edge case where onTimeout() was triggered
+        // *before* the call to connection.write had returned.
+        connectionHandler.startTimeout();
+
+        // Note that we could have already received the response here (so onSet() / onException() would have been called). This is
+        // why we only test for CANCELLED_WHILE_IN_PROGRESS below.
+
+        // If cancel() was called after we set the state to "in progress", but before connection.write had completed, it might have
+        // missed the new value of connectionHandler. So make sure that cancelHandler() gets called here (we might call it twice,
+        // but it knows how to deal with it).
+        if (queryStateRef.get() == QueryState.CANCELLED_WHILE_IN_PROGRESS)
+            connectionHandler.cancelHandler();
+    }
+
     private void logError(InetSocketAddress address, Throwable exception) {
         logger.debug("Error querying {}, trying next host (error is: {})", address, exception.toString());
         if (errors == null)
@@ -162,8 +193,6 @@ class RequestHandler implements Connection.ResponseCallback {
     }
 
     private void retry(final boolean retryCurrent, ConsistencyLevel newConsistencyLevel) {
-        queryStateRef.set(queryStateRef.get().startNext());
-
         final Host h = current;
         this.retryConsistencyLevel = newConsistencyLevel;
 
@@ -171,6 +200,8 @@ class RequestHandler implements Connection.ResponseCallback {
         manager.executor().execute(new Runnable() {
             @Override
             public void run() {
+                if (queryStateRef.get().isCancelled())
+                    return;
                 try {
                     if (retryCurrent) {
                         if (query(h))
@@ -185,9 +216,22 @@ class RequestHandler implements Connection.ResponseCallback {
     }
 
     public void cancel() {
-        isCanceled = true;
-        if (connectionHandler != null)
-            connectionHandler.cancelHandler();
+        // Atomically set a special QueryState, that will cause any further operation to abort.
+        // We want to remember whether a request was in progress when we did this, so there are two cancel states.
+        while (true) {
+            QueryState previous = queryStateRef.get();
+            if (previous.isCancelled()) {
+                return;
+            } else if (previous.inProgress && queryStateRef.compareAndSet(previous, QueryState.CANCELLED_WHILE_IN_PROGRESS)) {
+                // The connectionHandler should be non-null, but we might miss the update if we're racing with write().
+                // If it's still null, this will be handled by re-checking queryStateRef at the end of write().
+                if (connectionHandler != null)
+                    connectionHandler.cancelHandler();
+                return;
+            } else if (!previous.inProgress && queryStateRef.compareAndSet(previous, QueryState.CANCELLED_WHILE_COMPLETE)) {
+                return;
+            }
+        }
     }
 
     @Override
@@ -272,10 +316,8 @@ class RequestHandler implements Connection.ResponseCallback {
         }
 
         Host queriedHost = current;
+        boolean releaseConnection = true;
         try {
-            if (connection instanceof PooledConnection)
-                ((PooledConnection)connection).release();
-
             switch (response.type) {
                 case RESULT:
                     setFinalResult(connection, response);
@@ -384,31 +426,23 @@ class RequestHandler implements Connection.ResponseCallback {
                                 return;
                             }
 
+                            String currentKeyspace = connection.keyspace();
+                            String prepareKeyspace = toPrepare.getQueryKeyspace();
+                            if (prepareKeyspace != null && (currentKeyspace == null || !currentKeyspace.equals(prepareKeyspace))) {
+                                // This shouldn't happen in normal use, because a user shouldn't try to execute
+                                // a prepared statement with the wrong keyspace set.
+                                // Fail fast (we can't change the keyspace to reprepare, because we're using a pooled connection
+                                // that's shared with other requests).
+                                throw new IllegalStateException(String.format("Statement was prepared on keyspace %s, can't execute it on %s (%s)",
+                                    toPrepare.getQueryKeyspace(), connection.keyspace(), toPrepare.getQueryString()));
+                            }
+
                             logger.info("Query {} is not prepared on {}, preparing before retrying executing. "
                                       + "Seeing this message a few times is fine, but seeing it a lot may be source of performance problems",
                                         toPrepare.getQueryString(), connection.address);
-                            String currentKeyspace = connection.keyspace();
-                            String prepareKeyspace = toPrepare.getQueryKeyspace();
-                            // This shouldn't happen in normal use, because a user shouldn't try to execute
-                            // a prepared statement with the wrong keyspace set. However, if it does, we'd rather
-                            // prepare the query correctly and let the query executing return a meaningful error message
-                            if (prepareKeyspace != null && (currentKeyspace == null || !currentKeyspace.equals(prepareKeyspace)))
-                            {
-                                logger.debug("Setting keyspace for prepared query to {}", prepareKeyspace);
-                                connection.setKeyspace(prepareKeyspace);
-                            }
 
-                            queryStateRef.set(queryStateRef.get().startNext());
-                            try {
-                                connection.write(prepareAndRetry(toPrepare.getQueryString()));
-                            } finally {
-                                // Always reset the previous keyspace if needed
-                                if (connection.keyspace() == null || !connection.keyspace().equals(currentKeyspace))
-                                {
-                                    logger.debug("Setting back keyspace post query preparation to {}", currentKeyspace);
-                                    connection.setKeyspace(currentKeyspace);
-                                }
-                            }
+                            releaseConnection = false; // we're reusing it for the prepare call
+                            write(connection, prepareAndRetry(toPrepare.getQueryString()));
                             // we're done for now, the prepareAndRetry callback will handle the rest
                             return;
                         default:
@@ -447,6 +481,9 @@ class RequestHandler implements Connection.ResponseCallback {
         } catch (Exception e) {
             setFinalException(connection, e);
         } finally {
+            if (releaseConnection && connection instanceof PooledConnection)
+                ((PooledConnection)connection).release();
+
             if (queriedHost != null)
                 manager.cluster.manager.reportLatency(queriedHost, latency);
         }
@@ -474,6 +511,9 @@ class RequestHandler implements Connection.ResponseCallback {
                                  retryCount, queryState, queryStateRef.get());
                     return;
                 }
+
+                if (connection instanceof PooledConnection)
+                    ((PooledConnection)connection).release();
 
                 // TODO should we check the response ?
                 switch (response.type) {
@@ -593,10 +633,13 @@ class RequestHandler implements Connection.ResponseCallback {
     }
 
     // This is used to prevent races between request completion (either success or error) and timeout.
-    // A retry is in progress once we have written the request to the connection and until we get back a response or a timeout.
+    // A retry is in progress once we have written the request to the connection and until we get back a response (see onSet
+    // or onException) or a timeout (see onTimeout).
     // The count increments on each retry.
     static class QueryState {
-        static QueryState INITIAL = new QueryState(0, true);
+        static final QueryState INITIAL = new QueryState(-1, false);
+        static final QueryState CANCELLED_WHILE_IN_PROGRESS = new QueryState(Integer.MIN_VALUE, false);
+        static final QueryState CANCELLED_WHILE_COMPLETE = new QueryState(Integer.MIN_VALUE + 1, false);
 
         final int retryCount;
         final boolean inProgress;
@@ -620,9 +663,13 @@ class RequestHandler implements Connection.ResponseCallback {
             return new QueryState(retryCount + 1, true);
         }
 
+        public boolean isCancelled() {
+            return this == CANCELLED_WHILE_IN_PROGRESS || this == CANCELLED_WHILE_COMPLETE;
+        }
+
         @Override
         public String toString() {
-            return String.format("QueryState(count=%d, inProgress=%s)", retryCount, inProgress);
+            return String.format("QueryState(count=%d, inProgress=%s, cancelled=%s)", retryCount, inProgress, isCancelled());
         }
     }
 }