+++ driver-core/src/test/java/com/datastax/driver/core/PoolingOptionsTest.java
@@ -48,8 +48,9 @@ public class PoolingOptionsTest {
                                .hasState(State.UP)
                                .isAtDistance(HostDistance.LOCAL);
             // Wait for the node to be up, because apparently on Jenkins it's still only ADDED when we reach this line
+            // Waiting for NEW_NODE_DELAY_SECONDS+1 allows the driver to create a connection pool and mark the node up
             assertThat(cluster).host(2)
-                               .comesUpWithin(120, SECONDS)
+                               .comesUpWithin(Cluster.NEW_NODE_DELAY_SECONDS+1, SECONDS)
                                .isAtDistance(HostDistance.LOCAL);
 
             // Bring host 3 up, its presence should be acknowledged but it should be ignored
@@ -63,13 +64,13 @@ public class PoolingOptionsTest {
                                .hasState(State.UP)
                                .isAtDistance(HostDistance.LOCAL);
             assertThat(cluster).host(3)
-                               .comesUpWithin(120, SECONDS)
+                               .comesUpWithin(Cluster.NEW_NODE_DELAY_SECONDS+1, SECONDS)
                                .isAtDistance(HostDistance.IGNORED);
             assertThat(session).hasNoPoolFor(3);
 
             // Kill host 2, host 3 should take its place
             ccm.stop(2);
-            TestUtils.waitFor(CCMBridge.ipOfNode(3), cluster, 120);
+            TestUtils.waitFor(CCMBridge.ipOfNode(3), cluster);
 
             assertThat(cluster).host(1)
                                .hasState(State.UP)